+++ src/memo_helpers/validation_memo.py
@@ -1,7 +1,9 @@
 import click
 
 
-def selection_notes_validation(folder, edit, delete, move, add, flist, search, remove):
+def selection_notes_validation(
+    folder, edit, delete, move, add, flist, search, remove, export
+):
     used_flags = {
         "folder": bool(folder),
         "edit": edit,
@@ -11,6 +13,7 @@ def selection_notes_validation(folder, edit, delete, move, add, flist, search, r
         "flist": flist,
         "search": search,
         "remove": remove,
+        "export": export,
     }
 
     if add and not folder:
@@ -23,9 +26,9 @@ def selection_notes_validation(folder, edit, delete, move, add, flist, search, r
             "--flist must be used alone. It cannot be combined with other flags or --folder."
         )
 
-    modifier_flags = ["edit", "delete", "move", "remove", "search"]
+    modifier_flags = ["edit", "delete", "move", "remove", "search", "export"]
     used_modifiers = [f for f in modifier_flags if used_flags[f]]
     if len(used_modifiers) > 1:
         raise click.UsageError(
-            "Only one of --edit, --delete, --move, --remove or search can be used at a time."
+            "Only one of --edit, --delete, --move, --remove , --export or search can be used at a time."
         )